LayoutTests/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T6-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.2_A8_T6

PASS Expected parsing failure

TEST COMPLETE
